\name{BayesRB-package}
\alias{BayesRB-package}
\alias{BRB-package}
\docType{package}
\title{
	A novel MCMC based polygenic genetic risk score algorithm for dichotomous traits
}
\description{
BayesRB is a Bayesian approach for SNP effect estimation and genetic risk prediction for dichotomous traits. It is an extension of BayesR proposed by Moser et al. BayesR method performs well on the SNP effect estimation and genetic risk prediction, but it is designed to be appied to the data with quantitative traits. BayesRB allows the dichotomous outcomes. It inherited the characteristics of unbiasedness, accuracy, sparseness, robustness and powerfulness. 
}
